Corrected the issue with styling, and moved the new test to UrlAuthorizationConfigurerTests.java
