--- test/src/test/java/org/springframework/security/test/web/configurers/UrlAuthorizationConfigurerAnonymousTest.java
@@ -1,51 +1,0 @@
-/*
- * Copyright 2002-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.security.test.web.configurers;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.configurers.UrlAuthorizationConfigurer;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@Configuration
-public class UrlAuthorizationConfigurerAnonymousTest {
-
-    @Test
-    public void emptyTest() {
-        // Simply allow the Config class configure HttpSecurity.
-        // There must be no errors.
-    }
-
-
-    @EnableWebSecurity
-    static class Config extends WebSecurityConfigurerAdapter {
-        @Override
-        public void configure(HttpSecurity http) throws Exception {
-            // @formatter:off
-            http
-                .apply(new UrlAuthorizationConfigurer<>(null)).getRegistry()
-                    .anyRequest().anonymous();
-            // @formatter:on
-        }
-    }
-
-}
